Add extra case for custom container names
